Component BtnGoogle
We add a component to SignInScreen and SignUpScreen for singin with google
